Introduce a Dockerfile for fast adoption and isolated build environment.
In this commit I add a Dockerfile and new Makefile target, `make
container`, that builds the container with the container name
`unikraft/tools:latest`.  In addition to this, I had to update the namespace of
the three packages within the repository to include the full path the
git repository, e.g.:

`tools/utils_toolchain` -> `github.com/unikraft/tools/utils_toolchain`

I did so as the repository itself appears to be hosted primarily on Github as
opposed to Xenbits.  This also follows good Go practices in aim of distributing
code.

The Dockerfile itself has been split into three seperate builds,
`local`, `build` and `binary` that serve as seperate environments for
developing unikraft/unicore tools program; building the binary and for
distribtion of the binary, respectively.