kraft:  Introduce YAML specification system.
This commit re-organizes the kraft cli program to work with a backwards-
compatibile YAML specification for Unikraft builds.  The specification
itself can be used simply as `kraft.yaml` within a repository in order
to use the Unikraft build system.

For now three core commands work with this specification: configure,
build and clean.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>