kraft/commands/configure: Validate architecture and platform.
This commit provides greater error handling for the configure-time
of an application by providing the ability to select the architecture
and platform.  By default, all Unikraft platforms and architectures
are listed in the configure command.  However, once enacted, the
logic to determine whether the particular application has been
programmed against this architecture or platform can be used to
further validate the configure claim.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>