kraft/commands/init: Intrduce init command.
The init subcommand is capable of creating a new blank project
by serializing a Project object with a specific version of the
Unikraft core library as well as with targetted architecture
and platform.

In addition to this, a template repository can be used to bootstrap
the application.  This works by simply copying the sources of an
existing application into the current working directory.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>