kraft: Introduce `volume` and `network` directives and increase stability.
This commit introduces two new directives to the kraft.yaml
specification such that the runtime of a unikraft application can be
altered.  This initiative finds itself placed within the realm of
increasing usability of unikernel applications.  In addition to this,
much of the runtime semantics, specifically faucets of file systems is
tied closely to the construction of a unikernel itself.  For this
reason, it is now possible to add:

```
volumes:
  <my-vol-name>:
    mountpoint /path/to/file.ext
    type: 9pfs|initrd|qcow2|raw
```

to kraft.yaml files.  During any construction, the filesystem itself
can also be created.  In the case of 9pfs, however, simply a location
can be mounted as a shared space.

Additional usage includes the ability to attach to a network now
largely typified by the following directive:

```
networks:
  <my-net-name>:
    dhcp: bool
    bridge_driver: brctl|more-to-come
```

This will allow a user to specify how the unikraft application attaches
to a network bridge on the host machine.

Additional improvements include reliability in the built-in caching
mechanism and correct curation of Kconfig options based on provided
library names under the `libraries` directive.

Future improvements are already known.  This includes:

 - Adding more network bridge drivers;
 - Correctly identifying Kconfig options for volume types (currently
   codede into the VolumesType Enum);
 - Correctly ideentifying and initialzing Kconfig options for particular
   networking drivers (currently unsupported).

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>