kraft: Provide specification file choice order.
This commit resolves errors that appear when there is both a kraft.yaml
and a kraft.yml file present in a directory by selecting one.  This
commit also fixes  errors that occur when a kraft.y(a)ml file is empty
with a helpful error message.  Finally, remove old artifacts of the
"deps.json" specication days.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>