package/debian: Support major Debian vendors
This commit adds functionality to support the construction of
debian archives (.deb) for major Debian vendors and their
distributions, including:

  * debian:sid
  * debian:buster
  * debian:stretch
  * ubuntu:bionic
  * ubuntu:xenial
  * ubuntu:trusty

To accomplish this, dh_virtualenv and virtualenv is used to
locally package site-packages and so Python PyPI packages within
the debian distribution itself.

Additional tools have been included to ease the creation of the
kraft (known as 'unikraft-tools') to ease the creation of the
debian package.  Namely, Dockerfile.pkg-deb provides a working
environment with all the necessary software for the top-level
make target "pkg-deb".  To  build a debian package, simply call:

  PKG_VENDOR=debian PKG_DISTRIBUTION=sid make pkg-deb

This will produce artifacts within dist/

Signed-off-by: Alexander Jung <alexander.jung@neclab.eu>