kraft/commands/lib/init: Generate unikraft libraries.
This commit introduces a new tool to automaticaly generate a
new Unikraft library.  Simply calling:

  $ kraft lib init

Will start the process of prompting the developer with questions
so as to generate a Unikraft-compatible library.  The library
itself is composed using deep integration with the cookiecutter
project so as to increase generalability and ease with future
additions to the template itself.

The library can be generated using either a source archive (.zip,
.tar.gz) or from a GitHub repository.

Signed-off-by: Alexander Jung <alexander.jung@neclab.eu>