Add the dependency analyser with the go toolchain to kraft.
This patch adds a new series of commands `devel` which allows
to interact with the go toolchain. For now, only the dependency
analyser has been added to the kraft tool.

The go runtime will be installed by a script if it is not
present on the host. In addition, the go toolchain will be
downloaded from a Github repo and then installed. Arguments
handling is then performed within the toolchain.

Signed-off-by: Gaulthier Gain <gaulthier.gain@uliege.be>