Add dependency analysis to kraft
This patch adds a new series of commands devel which allows
to interact with the go toolchain. For now, only the dependency
analyser has been added to the kraft tool.

The go binaries are directly integrated to the kraft repo which
allows to avoid the setup and the installation of the go runtime.

Signed-off-by: Gaulthier Gain <gaulthier.gain@uliege.be>