manifest: Introduce Manifest
This commit introduces the concept of a remote manifest.  The remote
manifest represents a way of pointing to some Unikraft source code,
whether this is the core or a microlibrary.  The Manifest contains
meta information about the component which can later be used in
conjunction with the automatic retrieval of microlibraries before
the construction of a unikernel itself.

Signed-off-by: Alexander Jung <alexander.jung@neclab.eu>