cmd/list/provider: Change order of Github version check
This commit changes the priority of the versions saved to the cache
by parsing the GitHub releases last, this will override any tags
which are found previously.  A release from Github is more likely
to have more detailed and better explanations of the state of the
version rather than the tag.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>