cmd/run: Correctly initialize the platform's runner
This commit improves the reliability of kraft's runner mechanism
and adapts it to the newer target-based approach which was
introduced with the v0.5 specification bump.  In addition, a new
method was introduced into the Application class, run, which
holds the core abstraction and thus allows the python module to be
in itself extensible.

Additionally, with multiple targets being possible for a single
application, some decision logic was introduced into the run
sub-command for kraft which auto-selects the correct one if there
is only one thing to run.  If not, a new utility has been added
which allows the user to correctly select the binary they wish
to run.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>