{app,config}: Propagate configure --use-version down to config
This commit pushes the serialization of the version override
(typically used in the `kraft configure` command) into the config
manager.  By moving the override closer to the reading of the file
than during the application instantiation-time; we can pre-emptively
be alerted for errors during this instantiation.  Alternatively,
where an override is provided and the configuration read will not
be properly represented by the current cache.  For example, when
a `kraft.yaml` file indicates a specific version for a library,
we can override this before the Library class is instantiated,
which in itself will check whether that version is available
via the manifest.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>