Fix gcc container building on arm and x86_64
Before this commit, the gcc container image couldn't be built on arm or x86_64,
generating a `file not found error`.
Replace `aarch64-linux-gnu` in Dockerfile.gcc with the GCC_PREFIX that is
architecture specific. This way, the problematic folder will always be found.
This should also allow `kraft build` to cross-compile images on any architecture.