Integrate the devel toolchain to the kraft tool.
This commit adds a new series of commands "devel" which allows
to interact with the go toolchain. These tools can either be built
locally (`go` must be installed) or via Docker (`make docker-gotools`).

Each tool of the toolchain can thus be launched via the following
command: kraft-devel-*.

Note: at this stage, the toolchain is still in beta.

Signed-off-by: Gaulthier Gain <gaulthier.gain@uliege.be>