Review `support` package to be sync with latest changes.
In this commit, we migrate the latest changes of the `support` package
to the kraft repo. We divided the `support` into three different
categories. The first one `binary_analyser` contains an example file
used for the binary analysis. The second one `test_files` contains
several tests used for the dynamic analysis. And finally the last one
`config_files` contains configuration files for the dynamic
analysis.

Signed-off-by: Gaulthier Gain <gaulthier.gain@uliege.be>