Add GCC and QEMU version parameters to docker-kraft-dev task
At this moment, the CI / CD pipeline is running  `make docker-kraft TARGET=kraft-dev`,
without setting `GCC_VERSION` and `QEMU_VERSION` arguments. This results in old
`docker-gcc` and `docker-qemu` images being used when copying files to `docker-kraft`.

Most of the time, this is not a problem, but when we create  PRs that bring changes
to the `docker-gcc` or `docker-qemu` tasks, we may get misleading failures.

This PR adds the missing version parameters to the `docker-kraft-dev` task (as we
do for the simple `docker-kraft`).

Signed-off-by: Răzvan Vîrtan <virtanrazvan@gmail.com>